+++ cypress/fixtures/extensions/extension-bar/filters.js
@@ -1,0 +1,5 @@
+const { addFilter } = require('govuk-prototype-kit').nunjucks
+
+addFilter('bar__link',
+  (content, url) => `<a href="${url || '#'}">${content}</a>`,
+  { renderAsHtml: true })