+++ index.js
@@ -1,0 +1,8 @@
+const { getFilter, addFilter } = require('./lib/filters/api')
+
+module.exports = {
+  nunjucks: {
+    getFilter,
+    addFilter
+  }
+}